:twisted_rightwards_arrows: Merging branches : Merged

+++ backend/airlingo/src/main/java/com/ssafy/airlingo/domain/matching/response/MatchingUserDto.java
@@ -49,5 +49,5 @@ public class MatchingUserDto {
 	private List<String> userInterestLanguages;
 
 	@Schema(description = "사용자 별점")
-	private int userRating;
+	private double userRating;
 }
